REFACTOR test classes into `Unit` & `Feature` directories
