+++ tests/Feature/ResolveModelNameTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Sfneal\Models\Tests;
+namespace Sfneal\Models\Tests\Feature;
 
 use Sfneal\Models\Actions\ResolveModelName;
 use Sfneal\Models\Tests\Assets\Models\CompanyPeople;
 use Sfneal\Models\Tests\Assets\Models\People;
+use Sfneal\Models\Tests\ModelTestCase;
 
 class ResolveModelNameTest extends ModelTestCase
 {
